patra_agent/requirements.txt
python-dotenv
langchain
langgraph
langchain-mcp-adapters
langchain-anthropic
langchain-ollama
langchain-community
langgraph_supervisor

neo4j~=5.11.0
numpy~=1.26.4
scipy~=1.12.0
scikit-learn~=1.5.1
pandas~=2.2.1
flask~=3.0.3
flask-restx~=1.3.0
requests==2.28.1